refactor simulation
